Fixed corrupted Answer entities with null values in all fields. - Use JDO transactions to get and put the Question entities because it has an embedded Answer entities, requiring multiple atomic get/put queries. - Add extra filtering to get rid of all Question that might be incomplete due to null Answer entities. For some reason, there are still about 0.1% of the cases that are such even with the transactions imposed. - Made all repositories function to close PersistenceManager opened in a finally block to avoid unintended changes to the datastore entities to be persisted.
Fixed style guides to be consistent with the codebase as done in commit c8fd1577acd64496176546cddcb7139b34e4c8dc.
Fixed typo in GOOGLE_ANALYTICS_ID and erroneous check of gaq defined.
Fixed typo in AnswerKind of INVALID vs INCORRECT in angularjs.
Removed deprecated AnswerKind in angularjs, and made sure checking of correct answer for silver question to be consistent with the endpoints.